CDAP-13391 log provisioner messages in program context
Modify the provisioner tasks to use a program logging context
instead of a system logging context in order to tie the logs to
the program run and not the system logs. Even though the provisioner
runs in the CDAP master, it's logs belong with the program run,
otherwise a provisioning error will cause a program run to fail
without any feedback given to the user.

Also adding a cluster end time to each run record. This is needed
because the log handler restricts the log query end time using
the end time in the run record if it exists. However, the end time
in the run record is the program run end, whereas the cluster end
time is usually minutes later. Adjusted the log handler to restrict
using the cluster end time, so that the deprovisioning logs will
be included in the program logs.